+++ src/backend/core/enums.py
@@ -28,9 +28,6 @@ UPDATED_AT = "updated_at"
 USERS = "users"
 GROUPS = "groups"
 
-RELEVANCE = "relevance"
-
-ORDER_BY_OPTIONS = (RELEVANCE, TITLE, CREATED_AT, UPDATED_AT, SIZE, REACH)
 SOURCE_FIELDS = (
     TITLE,
     CONTENT,